requirements.txt
openai
gpt_index
python-dotenv
